Refactored creation of request to "Helper"
It's generic enough to be used in other samples.
